Record modes are hard

+++ tests/integration/helper.py
@@ -8,7 +8,9 @@ class IntegrationHelper(unittest.TestCase):
     def setUp(self):
         self.cassette_path = None
         self.session = Session()
+        self.cassette_created = True
 
     def tearDown(self):
-        assert self.cassette_path is not None
-        os.unlink(self.cassette_path)
+        if self.cassette_created:
+            assert self.cassette_path is not None
+            os.unlink(self.cassette_path)
